+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `fab`
   - Added `--mdc-fab-focus-outline-color`
   - Added `--mdc-fab-focus-outline-width`
+- `select`
+  - Forwarded `layoutOptions` method from foundation
 
 ## [v0.20.0] - 2020-12-03